### 2024-09-13 #7221 @editorialbot
[PRE REVIEW]: Psymple: A Python package for complex systems modelling
**Submitting author:** @GJHSimmons (<a href="http://orcid.org/0000-0002-9076-4973">George Simmons</a>)
**Repository:** https://github.com/casasglobal-org/psymple/
**Branch with paper.md** (empty if default branch): joss-review
**Version:** v0.3.4
