### 2024-09-13 #7221 @danielskatz
Also, JOSS papers typically don't use addresses in the affiliations, just institutions and countries.
