### 2024-09-13 #7221 @danielskatz
Finally, your paper is a bit long. JOSS suggests 250-1000 words, and your paper is a little over 1600. Is there anything you can remove, perhaps pointing to documentation or the repository instead?
