+++ paper.bib
@@ -16,7 +16,7 @@ doi = {10.1016/B978-0-12-809633-8.02390-6}
    volume = "40",
    number = "Volume 40, 2009",
    pages = "677-697",
-   doi = "https://doi.org/10.1146/annurev.ecolsys.110308.120159",
+   doi = "10.1146/annurev.ecolsys.110308.120159",
    url = "https://www.annualreviews.org/content/journals/10.1146/annurev.ecolsys.110308.120159",
    publisher = "Annual Reviews",
    issn = "1545-2069",
@@ -38,7 +38,7 @@ volume = {12},
 number = {4},
 pages = {334-350},
 keywords = {Biophysical ecology, climate, fundamental niche, geographical information systems, microclimate, physiological ecology, realized niche, species distribution modelling},
-doi = {https://doi.org/10.1111/j.1461-0248.2008.01277.x},
+doi = {10.1111/j.1461-0248.2008.01277.x},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1461-0248.2008.01277.x},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1461-0248.2008.01277.x},
 abstract = {Abstract Species distribution models (SDMs) use spatial environmental data to make inferences on species’ range limits and habitat suitability. Conceptually, these models aim to determine and map components of a species’ ecological niche through space and time, and they have become important tools in pure and applied ecology and evolutionary biology. Most approaches are correlative in that they statistically link spatial data to species distribution records. An alternative strategy is to explicitly incorporate the mechanistic links between the functional traits of organisms and their environments into SDMs. Here, we review how the principles of biophysical ecology can be used to link spatial data to the physiological responses and constraints of organisms. This provides a mechanistic view of the fundamental niche which can then be mapped to the landscape to infer range constraints. We show how physiologically based SDMs can be developed for different organisms in different environmental contexts. Mechanistic SDMs have different strengths and weaknesses to correlative approaches, and there are many exciting and unexplored prospects for integrating the two approaches. As physiological knowledge becomes better integrated into SDMs, we will make more robust predictions of range shifts in novel or non-equilibrium contexts such as invasions, translocations, climate change and evolutionary shifts.},
@@ -69,7 +69,7 @@ volume = {18},
 number = {1},
 pages = {93-109},
 keywords = {climate change, ecological niche model, integrative model, species distribution model, vertebrates},
-doi = {https://doi.org/10.1111/1749-4877.12618},
+doi = {10.1111/1749-4877.12618},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/1749-4877.12618},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/1749-4877.12618},
[…]
