+++ paper.bib
@@ -270,7 +270,7 @@ doi = {10.1079/9781800621459.0015}
   title={Simulation of ecological processes},
   edition={2},
   year={1978},
-  doi={10.1016/0304-3800(80)90048-4}
+  doi={10.1016/0304-3800(80)90048-4},
   publisher={Centre for Agricultural Publishing and Documentation, Wageningen, The Netherlands}
 }
 
