+++ paper.bib
@@ -242,7 +242,7 @@ title = {Analysis of invasive insects: Links to climate change},
 isbn = {978-1-80062-143-5},
 journal = {Invasive Species and Global Climate Change},
 edition = {Second},
-doi = {10.1079/9781800621459.0004}
+doi = {10.1079/9781800621459.0000}
 }
 
 @inbook{gu-po:2022b,
@@ -254,7 +254,7 @@ title = {Assessing and Managing the Impact of Climate Change on an Invasive Weed
 isbn = {978-1-80062-143-5},
 journal = {Invasive Species and Global Climate Change},
 edition = {Second},
-doi = {10.1079/9781800621459.0015}
+doi = {10.1079/9781800621459.0000}
 }
 
 @inproceedings{fi-ni:1970,
