### 2024-09-16 #7221 @GJHSimmons
@danielskatz Thank you for proceeding quickly.

I've fixed DOIs and affiliations. Flagging that two chapters ([here](https://www.cabidigitallibrary.org/doi/10.1079/9781800621459.0004) and [here]https://www.cabidigitallibrary.org/doi/10.1079/9781800621459.0015) have stated but inactivated DOIs. I've fallen back to the [full book](doi.org/10.1079/9781800621459.0000) which has an active DOI.

I'll need some deeper advice on paper length. I understand the 'description' section is unusual, but contains important citations for how the package was built which don't belong in the statement of need or documentation. How is this usually handled for JOSS articles?
