### 2024-10-17 #7364 @editorialbot
[REVIEW]: Psymple: A Python package for complex systems modelling
**Submitting author:** @GJHSimmons (<a href="http://orcid.org/0000-0002-9076-4973">George Simmons</a>)
**Repository:** https://github.com/casasglobal-org/psymple/
**Branch with paper.md** (empty if default branch): joss-review
**Version:** 1.0.2
