### 2024-10-18 casasglobal-org/psymple#48 @SunnyXu
Contribution and authorship
If my observation is correct, there are only two authors committed to the GitHub Code. However, there are six authors listed in the article. Maybe a statement of contribution in the documentation or somewhere is necessary to avoid potential confusion.
