### 2024-10-18 casasglobal-org/psymple#49 @SunnyXu
A statement of need in the Documentation
Maybe you have mentioned partially the statement of need somewhere in the documentation. However, it would be good to create a section in the documentation or readme to highlight it.
