### 2024-10-18 casasglobal-org/psymple#50 @SunnyXu
Automated tests
I see there is a folder of tests on your Github web. However, there seem no automated tests or manual steps described so that the functionality of the software can be verified. Please add it on.
