### 2024-10-18 casasglobal-org/psymple#51 @SunnyXu
Community Guidelines
There seem no clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. Please add it on.
