### 2024-10-18 casasglobal-org/psymple#52 @SunnyXu
Word counts in manuscript
The word count for paper.md is 1594 right now. However, according to JOSS Paper Format (https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain
), please consider shortening the manuscript a bit if possible.
