### 2024-10-18 casasglobal-org/psymple#53 @SunnyXu
Break links in readme
Links in the “Where to start” section in the readme are all broken as 404. Please revise.
