### 2024-10-18 casasglobal-org/psymple#54 @SunnyXu
Typo in the example of multiple tanks
Maybe at the end of the multiple tanks (https://casasglobal-org.github.io/psymple/0.3.4/examples/mixing_problems/multiple_tanks/), the ““rate_out_2”: 0” is not supposed to be commanded out? Please double check.
