### 2024-10-18 casasglobal-org/psymple#55 @SunnyXu
Add more examples/tutorials if possible
It would be good to include more examples/tutorials with figures in the documentation, which might be more attractive. More examples would be very helpful too.
