### 2024-10-18 #7364 @SunnyXu
@prashjha  @GJHSimmons

Psymple is a solid Python package designed to facilitate the development of hybrid complex system models and modeling frameworks, that apply to diverse systems. The paper is written clearly and easy to follow. Previous packages and related articles are well referred to. However, please see my comments for possible improvements.

Major points:
(1) Contribution and authorship (https://github.com/casasglobal-org/psymple/issues/48)

(2) A statement of need in the Documentation (https://github.com/casasglobal-org/psymple/issues/49)

(3) Automated tests (https://github.com/casasglobal-org/psymple/issues/50)

(4) Community Guidelines (https://github.com/casasglobal-org/psymple/issues/51)

(5) Word counts in the manuscript (https://github.com/casasglobal-org/psymple/issues/52)

Minor points:
(1) Break links in the readme (https://github.com/casasglobal-org/psymple/issues/53)

(2) Typo in the example of multiple tanks (https://github.com/casasglobal-org/psymple/issues/54)

(3) Add more examples/tutorials if possible (https://github.com/casasglobal-org/psymple/issues/55)
