### 2024-10-18 8ca290e
Update variables_parameters.md
