### 2024-10-18 a12a540
Update multiple_tanks.md
