### 2024-10-18 casasglobal-org/psymple#57 @pitmonticone
[REVIEW]: Missing DOIs
Dear authors,

Please review the [missing DOIs](https://github.com/openjournals/joss-reviews/issues/7364#issuecomment-2420703213) and ensure the references are updated accordingly.

Once you have completed this, please confirm by posting the following comment in the review thread [here](https://github.com/openjournals/joss-reviews/issues/7364):

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7364
