### 2024-10-19 casasglobal-org/psymple#59 @pitmonticone
[REVIEW]: SyntaxWarning in `1-single_tank.py` at lines 12 and 89
When running the example script `examples/mixing_problems/1-single_tank.py` I encountered the following warning at lines 12 and 89:

```console
SyntaxWarning: invalid escape sequence '\('
```

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7364
