### 2024-10-19 casasglobal-org/psymple#60 @pitmonticone
[REVIEW]: SyntaxWarning in `2-multiple_tanks.py` at line 15
When running the example script `examples/mixing_problems/2-multiple_tanks.py` I encountered the following warning at line 15:

```console
SyntaxWarning: invalid escape sequence '\('
```

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7364
