### 2024-10-19 casasglobal-org/psymple#61 @pitmonticone
[REVIEW]: SyntaxWarning in `1-malthusian_population.py` at line 13
When running the example script `examples/population_dynamics/1-malthusian_population.py` I encountered the following warning at line 13:

```console
SyntaxWarning: invalid escape sequence '\('
```

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7364
