### 2024-10-19 casasglobal-org/psymple#62 @pitmonticone
[REVIEW]: UserWarning in `1-projectile_motion.py`
When running the example script `examples/second_order_ODEs/1-projectile_motion.py` I encountered the following warnings

```console
UserWarning: A signature for function g was not provided. The appearance of the function in displayed expressions may not be as expected.
```

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7364
