### 2024-10-19 casasglobal-org/psymple#63 @pitmonticone
[REVIEW]: UserWarning in `2-three_body_problem.py`
When running the example script `examples/second_order_ODEs/2-three_body_problem.py` I encountered the following warnings

```console
UserWarning: A signature for function G was not provided. The appearance of the function in displayed expressions may not be as expected
```

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7364
