### 2024-10-25 casasglobal-org/psymple#57 @GJHSimmons
@pitmonticone Firstly, thank you for agreeing to review this package. I will be working through the issues properly next week.

The three 'skipped' have no corresponding DOIs (2x are old papers with no online version, 1x is a Github repo which has not had a DOI created). Please advise how to continue.
