### 2024-10-25 casasglobal-org/psymple#50 @GJHSimmons
@SunnyXu Firstly, thank you for reviewing this package. I will be properly working through your comments next week.

Tests are run automatically on pushes to main branch using [test.yml](https://github.com/casasglobal-org/psymple/blob/main/.github/workflows/test.yml). Please could you clarify your request?
