### 2024-10-26 casasglobal-org/psymple#50 @SunnyXu
@GJHSimmons
Hi George,
Sure. Happy to reveiw your work.
Sorry, but I did not notice the test.yml file at the first glance.
It would be good to add a section in the readme on GitHub to notify the community regarding the automated tests. If possible, a automated code coverage on GitHub would convince the readers that your package is reliable.
Please take your time to revise the manuscript and the package and leave a comment to let me know once you are done.
Have a good one.
Jin
