### 2024-10-27 casasglobal-org/psymple#57 @pitmonticone
Hi @GJHSimmons, thank you for your response.

As there are no corresponding DOIs, I believe this resolves the issue.
