+++ docs/examples/second_order_odes/three_body_problem.md
@@ -1,0 +1,249 @@
+# Three body problem
+
+The three body problem, or more generally the $n$-body problem, is a classical problem in celestial mechanics. The statement of the problem is to find the trajectories of three or more massive bodies under the gravitational influence of each other. 
+
+## System of equations
+
+The magnitude of the gravitational force acting between two objects of masses $m_1$ and $m_2$ is given by $F = \frac{Gm_1m_2}{r^2}$, where $G = \pu{6.67 \times 10^{-11} N m^2/kg}$ and $r$ is the distance between their centres. 
+
+For the dynamics of two or more bodies, each body experiences a gravitational force from every other body in the direction of that body with the above magnitude. Therefore for $n$ bodies, the system of differential equations is derived by applying Newton's second law:
+
+$$
+\frac{d^2 \mathbf{r}_i}{dt^2} = \sum_{1 \leqslant i \ne j \leqslant n} G m_j \frac{(\mathbf{r}_j - \mathbf{r}_i)}{|\mathbf{r}_j - \mathbf{r}_i|^3}, \qquad i=1,\dots,n.
+$$
+
+## Setup in `psymple`
+
+To model the $n$-body problem in `psymple`, for simplicity assume the motion is happening in a plane, so that $\mathbf{r}_i = (x_i, y_i)$. Then each differential equation above corresponds to a pair of second-order differential equations in $x_i$ and $y_i$. To implement this in `psymple` requires an additional pair of equations for $v_{x_i}$ and $v_{y_i}$, the components of the velocity $\frac{d \mathbf{r}_i}{dt}$. 
+
+The following implementation models the force components
+
+$$
+F_{ij} = \frac{Gm_j}{((x_j-x_i)^2 + (y_j - y_i)^2)^{\frac{3}{2}}} \left(x_j - x_i, y_j-y_i \right)
+$$
+
+for $i \ne j$. The velocity components of body $i$ satisfy the differential equations obtained by aggregating the components of $F_{ij}$ for $j=1,\dots,i-1,i+1,\dots,n$. 
+
+### Common objects
+
+The following objects are common calculations for each $F_{ij}$. 
+
+```py
+from psymple.build import VariablePortedObject, FunctionalPortedObject
+
+vars = VariablePortedObject(
+    name="vars",
+    assignments=[
+        ("v_x", "mu * Del_x"),
+        ("v_y", "mu * Del_y"),
+    ]
+)
+
+dist = FunctionalPortedObject(
+    name="dist",
+    assignments=[
+        ("Del_x", "x_o - x"),
+        ("Del_y", "y_o - y"),
+        ("d", "sqrt((x_o-x)**2 + (y_o-y)**2)"),
+    ]
+)
+
+force = FunctionalPortedObject(
+    name=f"force",
+    assignments=[("mu", "G*m/d**3")]
+)
+```
+
+### Velocity model
+
+The following class `velocity` is used to create the model for $F_{ij}$, with the argument `id` used to identify each pair $(i,j)$. 
+
+```py
+from psymple.build import CompositePortedObject 
+
+class velocity(CompositePortedObject):
+
+    def __init__(self, id):
+        super().__init__(
+            name=f"velocity_{id}",
+            children=[vars, force, dist],
+            input_ports=["x", "y", "x_o", "y_o", "m_o"],
+            variable_ports=["v_x", "v_y"],
+            directed_wires=[
+                ("m_o", "force.m"),
+                ("x", "dist.x"),
+                ("y", "dist.y"),
+                ("x_o", "dist.x_o"),
+                ("y_o", "dist.y_o"),
+                ("dist.Del_x", "vars.Del_x"),
+                ("dist.Del_y", "vars.Del_y"),
+                ("dist.d", "force.d"),
[…]
