+++ docs/user_guide/getting_started.md
@@ -1,8 +1,12 @@
-# Getting started
+# Introduction
 
-`psymple` is a Python package currently supporting versions `3.10` and higher. 
+`psymple` is a package designed for general complex systems modelling. A complex system is a dynamic system whose interconnectivity produces non-trivial or unexpected behaviour. Examples of complex systems exist everywhere, from ecological systems, chemical reactions, economic systems, astronomy and even social networks.
+
+The [sections below](#where-to-start) give a wide introduction to the structure of `psymple`, why it exists, how to use it, and how to get involved.
 
 ## Installation
+
+`psymple` is a Python package currently supporting versions `3.10` and higher. 
 
 The latest version of `psymple` is available on [PyPI](https://pypi.org/project/psymple/) and can be installed with `pip`.
 
@@ -12,8 +16,56 @@ pip install psymple
 
 ## Where to start
 
-Familiarisation with the purpose and design principles of `psymple` is recommended before beginning. You can start reading these here: [hybrid systems modelling](../overview/modelling_systems.md).
+For most new users, going through the following resources in the suggested order will provide a good introduction to `psymple`. 
 
-Once you're happy, you can start modelling! The best pages to start are [functions](../components/functional_ported_objects.md), [differential equations](../components/variable_ported_objects.md) and [composite systems](../components/composite_ported_objects.md).
+<div class="grid cards" markdown>
 
-If you're interested in how `psymple` implements mathematical principles into `python` objects, the best place to start is [ported objects](../mathematics/ported_objects.md).
+-   :material-roman-numeral-1:{ .lg .middle } __Why psymple exists__
+
+    ---
+
+    Find out about the purpose and design principles of `psymple`.
+
+    [:octicons-arrow-right-24: Read more](../overview/introduction.md)
+
+-   :material-roman-numeral-2:{ .lg .middle } __Modelling in psymple__
+
+    ---
+
+    Learn how to capture functions and differential equations in `psymple`, and combine these into complex models.
+
+    [:octicons-arrow-right-24: Start modelling](../components/functional_ported_objects.md)
+
+-   :material-roman-numeral-3:{ .lg .middle } __Simulating models__
+
+    ---
+
+    Understand how to create and run simulations for `psymple` models.
+
+    [:octicons-arrow-right-24: Run simulations](./system.md)
+
+-   :material-roman-numeral-4:{ .lg .middle } __Tutorials__
+
+    ---
+
+    Work through full examples of implementing, parametrising and simulating models.
+
+    [:octicons-arrow-right-24: See more](../examples/tutorials.md)
+
+-   :material-roman-numeral-5:{ .lg .middle } __The mathematics__
+
+    ---
+
+    Learn more about the mathematical structures underneath psymple which tie everything together.
+
+    [:octicons-arrow-right-24: See the maths](../mathematics/ported_objects.md)
+
+-   :material-roman-numeral-6:{ .lg .middle } __Get involved__
+
+    ---
+
+    Find out how you can get involved in the community by using or developing `psymple`.
+
+    [:octicons-arrow-right-24: Get involved](../development/development.md)
+
+</div>
