+++ mkdocs.yml
@@ -15,6 +15,7 @@ theme:
     - content.code.annotate
     - navigation.tabs
     - navigation.tabs.sticky
+    - navigation.indexes
     - navigation.sections
     - navigation.footer
     - search.suggest
@@ -69,8 +70,14 @@ extra:
     default: latest
 
 nav:
+  - Home: index.md
+  - Overview:
+    - Introduction: overview/introduction.md
+    - Statement of need: overview/statement_of_need.md
+    - Hybrid systems modelling: overview/modelling_systems.md
+    - Mechanistic and correlative models: overview/modelling_types.md
+    - System design: overview/system_design.md
   - User guide:
-    - user_guide/getting_started.md
     - Building a model:
       - Defining functions: components/functional_ported_objects.md
       - Defining ODEs: components/variable_ported_objects.md
@@ -80,25 +87,19 @@ nav:
       - Building a system: user_guide/system.md
     - Running a simulation: 
       - Simulations: user_guide/simulation.md
-  - Overview:
-    - Introduction: overview/introduction.md
-    - Statement of need: overview/statement_of_need.md
-    - Hybrid systems modelling: overview/modelling_systems.md
-    - Mechanistic and correlative models: overview/modelling_types.md
-    - System design: overview/system_design.md
   - Examples:
-    - examples/tutorials.md
+    - Examples: examples/index.md
     - Population dynamics:
-      - Population modelling: examples/population_dynamics/population_modelling.md
+      - Population modelling: examples/population_dynamics/index.md
       - Malthusian growth: examples/population_dynamics/malthusian_population.md
       - Logistic growth: examples/population_dynamics/logistic_growth.md
       - Predator-prey systems: examples/population_dynamics/predator_prey.md
     - Mixing problems: 
-      - Mixing problems: examples/mixing_problems/mixing_problems.md
+      - Mixing problems: examples/mixing_problems/index.md
       - Single tank: examples/mixing_problems/single_tank.md
       - Multiple tanks: examples/mixing_problems/multiple_tanks.md
     - Second-order ODEs:
-      - Second-order ODEs: examples\second_order_odes\second_order_odes.md
+      - Second-order ODEs: examples\second_order_odes\index.md
       - Projectile motion: examples\second_order_odes\projectile_motion.md
       - Three body problem: examples\second_order_odes\three_body_problem.md
   - Mathematical components:
