### 2025-01-22 casasglobal-org/psymple#55 @GJHSimmons
Thank you for your patience, I've now had time to go through and address this.

Please review the updated set of examples in the documentation [here](https://casasglobal-org.github.io/psymple/dev/examples/) which have been organised and contain figures.
