### 2025-01-22 casasglobal-org/psymple#49 @GJHSimmons
The [overview section](https://casasglobal-org.github.io/psymple/dev/overview/introduction/) now contains all the introduction to the package, as well as the full [statement of need](https://casasglobal-org.github.io/psymple/dev/overview/statement_of_need/)
