### 2025-01-22 casasglobal-org/psymple#48 @GJHSimmons
The [development team page](https://casasglobal-org.github.io/psymple/dev/development/development_team/) now gives information about all the listed authors and their contributions under the "Core governance and steering team" section. Please let us know if more clarity is needed.
