### 2025-01-22 casasglobal-org/psymple#54 @GJHSimmons
Thank you for spotting this, this has now been fixed in the [latest version](https://casasglobal-org.github.io/psymple/dev/examples/mixing_problems/multiple_tanks/)
