### 2025-01-22 casasglobal-org/psymple#53 @GJHSimmons
The links in the README should now be working again.
