+++ examples/population_dynamics/2-logistic_population.py
@@ -5,13 +5,12 @@ sys.path.insert(0, str(pathlib.Path(__file__).parents[2]))
 
 from psymple.build import (
     CompositePortedObject,
-    FunctionalPortedObject,
     VariablePortedObject,
     System,
 )
 
 """
-Recall the following variable object 'pop' capturing Malthusian growth in a reusable way.
+Basic population model
 """
 
 pop = VariablePortedObject(
@@ -21,131 +20,65 @@ pop = VariablePortedObject(
 )
 
 """
-Logistic population growth consists of a Malthusian growth component and a density-dependent limit to some
-carrying capacity K. The differential equation is dx/dt = rx(1 - x/K). 
-
-We could create a new variable object capturing this equation, but since we have already have the object
-'pop' capturing Malthusian growth, it feels as though we are repeating some work. 
-
-Here is an alternative way of viewing the logistic growth equation: its right-hand side is the sum of 
-two components: a Malthusian component rx, and a limit component -x^2/K. Psymple allows us to combine
-two variable objects, each capturing one of these components, to a single dynamic model in a process
-called composition. 
-
-In simple terms, the composition of the equations dx/dt = f(x,t) and dx/dt = g(x,t) is the equation
-dx/dt = f(x,t) + g(x,t). Let's see how to do this in psymple. First we need a variable object 
-capturing the logistic component, with a default value of 'K = 10'.
+Direct logistic implementation
 """
 
-limit = VariablePortedObject(
-    name="logistic_limit",
-    assignments=[("x", "- r* x**2 / K")],
+logistic_simple = VariablePortedObject(
+    name="logistic_simple",
+    assignments=[("x", "r*x*(1-x/K)")],
     input_ports=[("r", 0.1), ("K", 10)],
 )
 
 """
-Next, in a composite object, we use a variable wire to tell psymple to compose the variables
-'malthusian_pop.x' and 'logistic_limit.x' together. To do this, we specify a wire
-`wire = {"child_ports": ["malthusian_pop.x", "logistic_limit.x"], "output_name" = "x"}`. There is
-a shorthand for this, which at first appears unnatural:
-`wire = (["malthusian_pop.x", "logistic_limit.x"], None, "x")`. We will discuss this shortly.
+Structural implementation
+"""
+
+limit = VariablePortedObject(
+    name="limit",
+    assignments=[("x", "- r/K* x**2")],
+    input_ports=[("r", 0.1), ("K", 10)],
+)
+
+logistic_pop = CompositePortedObject(
+    name="logistic_pop",
+    children=[pop, limit],
+    variable_ports = ["x"],
+    variable_wires=[(["malthusian_pop.x", "limit.x"], "x")],
+)
+
+"""
+Sharing input parameters
 """
 
 logistic_pop = CompositePortedObject(
     name="logistic_pop",
     children=[pop, limit],
-    variable_wires=[(["malthusian_pop.x", "logistic_limit.x"], None, "x")],
-)
-
-system = System(logistic_pop)
-system.compile()
-
-sim = system.create_simulation("sim", solver="discrete", initial_values={"x": 1})
-sim.simulate(100, n_steps=10)
-sim.plot_solution()
-
-"""
-You'll that the simulation looks exactly as expected: the population initially undergoes exponential
-growth, before its growth rate slows and it approaches the limit of 10. 
-
-Let's discuss some efficiencies and good practice. Suppose instead that we did implement the logistic
-growth equation as a single variable object:
-
-logistic_pop = VariablePortedObject(
-    name="logistic_pop",
-    assignments=[("x", "r*(x - x**2/K)")],
     input_ports=[("r", 0.1), ("K", 10)],
-)
-
-As discussed, this implementation does not exploit the reusability features of psymple, but there are three
-apparent advantages:
[…]
