### 2025-01-22 8a6b684
mixing problem examples update
