+++ examples/mixing_problems/1-single_tank.py
@@ -9,37 +9,6 @@ from psymple.build import (
     System,
 )
 
-"""
-In these examples we will consider how to model _mixing problems_ in `psymple`. A mixing problem
-seeks to understand the evolution of a quantity of solvent in a solute, for example a chemical
-being mixed in a tank of water. These problems can be highly varied, for example solution can
-enter a tank at variable concentration or rate, or leave the tank at variable rate. Alternatively,
-several tanks can be connected together, with their solutions being pumped into each other.
-
-We will assume that tanks are _well-mixed_, that is, the concentration of the solute is the same
-throughout the tank at any moment in time. 
-
-First, we will consider a single tank with an initial volume of water \( V_0 l \) and an initial
-amount of \( M_0 g \) of salt dissolved in it. A solution with concentration \( c(t) g/l \) of
-salt flows into the tank at rate \( r_0(t) l/s \) and the mixed solution flows out of the tank at
-a rate of \( r_1(t) l/s \).
-
-Let \( V(t) \) be the volume of solution in the tank at time \( t \). Then 
-\( V'(t) = r_0(t) - r_1(t) \). Furthermore, let \( M(t) \) be the amount of salt in the solution
-at time \( t \). The rate of change of salt in the solution is given by
-\( M'(t) = r_0(t) c(t) - r_1(t) M(t)/V(t) \).
-
-In `psymple`, we could model \( V(t) \) and \( M(t) \) by writing just writing the two differential
-equations above. In that case, however, our model would become fixed, and if we considered a new
-situation with more than one in-flow or out-flow, we would have to build something new. Instead,
-consider what is happening at each point a pipe meets the tank. For the in-flow pipe, there are
-fluxes \( V'(t) = r_0(t) \) and \( M'(t) = r_0 (t) c(t) \). For the out-flow pipe, we similarly
-get \( V'(t) = -r_1(t) \) and \( M'(t) = - r_1(t) M(t)/V(t) \). The situation inside the tank
-is then nothing but the aggregation of these volume and mass variables.
-
-We model this initially with two `VariablePortedObject` instances, as follows.
-"""
-
 pipe_in = VariablePortedObject(
     name="pipe_in",
     assignments=[
@@ -51,20 +20,15 @@ pipe_in = VariablePortedObject(
 pipe_out = VariablePortedObject(
     name="pipe_out",
     assignments=[
-        ("V", "-r_1"),
+        ("V", "Piecewise((-r_1, V>V_m), (0, True))"), # (1)!
         ("M", "-r_1 * M/V"),
     ]
 )
 
-"""
-Using a `CompositePortedObject`, we can then aggregate the variables of these two objects together,
-and expose the parameters `r_0`, `r_1` and `c`.
-"""
-
 tank = CompositePortedObject(
     name="tank",
     children=[pipe_in, pipe_out],
-    input_ports=["r_0", "r_1", "c"],
+    input_ports=["r_0", "r_1", "c", "V_m"],
     variable_ports=["V", "M"],
     variable_wires=[
         (["pipe_in.V", "pipe_out.V"], "V"),
@@ -73,44 +37,29 @@ tank = CompositePortedObject(
     directed_wires=[
         ("r_0", "pipe_in.r_0"),
         ("r_1", "pipe_out.r_1"),
-        ("c", "pipe_in.c")
+        ("c", "pipe_in.c"),
+        ("V_m", "pipe_out.V_m"),
     ]
 )
-
-"""
[…]
