+++ examples/mixing_problems/2-multiple_tanks.py
@@ -7,20 +7,12 @@ sys.path.insert(0, str(pathlib.Path(__file__).parents[2]))
 
 from psymple.build import (
     CompositePortedObject,
-    FunctionalPortedObject,
     VariablePortedObject,
     System,
 )
 
 """
-In the previous example, we considered a mixing problem with a single tank, which had an external 
-flow in and an external flow out. Now we consider a mixing problem consisting of two or more
-tanks, each of which can have external flows in and out in the same manner as for the single tank,
-and which furthermore can have flows to and from other tanks in the system.
-
-It will help to first specify the system data. In this case, we will create a system with \( n > 1 \)
-tanks, such that every tank has a flow in and flow out, plus a flow to and from every other tank.
-The following data can be freely altered to reduce the external flows and connections, if desired.
+System data
 """
 
 # Number of tanks
@@ -39,9 +31,7 @@ pipes_in_id = [1, 2]
 pipes_out_id = [1, 2]
 
 """
-If you followed through the single tank example, the following lists of pipes in and out are created
-in exactly the same way as there. We create pipes in for every tank in `pipes_in_id` and pipes out
-for every tank in `pipes_out_id`.
+Flow building blocks
 """
 
 pipes_in = [
@@ -58,22 +48,13 @@ pipes_in = [
 pipes_out = [
     VariablePortedObject(
         name=f"out_{i}",
-        assignments=[("Q_out", "-rate * Q_out / V_out"), ("V_out", "-rate")],
+        assignments=[
+            ("Q_out", "-rate * Q_out / V_out"), 
+            ("V_out", "-rate")
+        ],
     )
     for i in pipes_out_id
 ]
-
-"""
-Next, we need to define connector pipes. These have four variables:
-
-- `Q_in`, the amount of salt entering the pipe from a tank,
-- `Q_out`, the amount of salt flowing into the next tank. Note that this has the same concentration
-    as for `Q_in`,
-- `V_in`, the volume of water entering the pipe from the tank,
-- `V_out`, the volume of water exiting the pipe into the next tank.
-
-We create a connector pipe for every pair in `link_pipes_id`.
-"""
 
 connectors = [
     VariablePortedObject(
@@ -89,40 +70,29 @@ connectors = [
 ]
 
 """
-Now we just need to define our system of tanks. Compared to the single tank example, 
-the children, variable ports and input ports are exactly the same (just one for each tank, 
-if it was specified), except for additionally the specification of the flow rates `rate_i_j` 
-of the connector pipes as additional input ports.
-
-The directed wires simply connect the rates at the input ports to the correct pipes.
-
-The only thing which becomes more complicated is the aggregation of the variables. Each tank `i`,
-and therefore each variable `Q_i` (and `V_i`), can have:
-
-- External in-flows if `i in pipes_in_id`,
-- External out-flows if `i in pipes_out_id`,
-- A flow to tank `j` for every `j` such that `(i,j) in link_pipes_id`,
-- A flow from tank `j` for every `j` such that `(j,i) in link_pipes_id`.
-
-The variable_wires data simply aggregates all these variables to `Q_i` (and `V_i`).
+Model of the tanks by aggregation
 """
 
 tanks = CompositePortedObject(
     name="tanks",
[…]
