+++ examples/second_order_ODEs/1-projectile_motion.py
@@ -12,6 +12,70 @@ from psymple.build import (
     System,
 )
 
+"""
+Setting up the system contextualisation
+"""
+
+frac_0 = lambda a,b,d: a/b if b != 0 else d
+
+system = System()
+system.add_system_parameter("g", "9.81"),
+system.add_system_parameter("rho", "1.225")
+system.add_utility_function(name="frac_0", function=frac_0)
+
+"""
+Vertical motion model
+"""
+
+v_gravity = VariablePortedObject( 
+    name="v_gravity",
+    assignments=[("v", "g"), ("x", "-v")], 
+)
+
+v_drag = VariablePortedObject(
+    name="v_drag",
+    assignments=[("v", "-mu * v**2")],
+)
+
+f_drag = FunctionalPortedObject(
+    name="f_drag",
+    assignments=[("mu", "frac_0(1/2 * C * rho * A, m, 0)")], 
+)
+
+model = CompositePortedObject(
+    name="model",
+    children=[v_gravity, v_drag, f_drag],
+    input_ports=["C", "A", "m"],
+    variable_ports=["v", "x"],
+    directed_wires=[
+        ("C", "f_drag.C"),
+        ("A", "f_drag.A"),
+        ("m", "f_drag.m"),
+        ("f_drag.mu", "v_drag.mu"), 
+    ],
+    variable_wires=[
+        (["v_gravity.v", "v_drag.v"], "v"),
+        (["v_gravity.x"], "x"),
+    ],
+)   
+
+"""
+Vertical motion simulation
+"""
+
+system.set_object(model)
+
+sim = system.create_simulation(
+    initial_values = {"v": 0, "x": 200}, 
+    input_parameters={"C": 1.1, "A": 0.2, "m": 2})
+sim.simulate(t_end=10)
+
+sim.plot_solution()
+
+"""
+Planar motion model
+"""
+
 motion = VariablePortedObject(
     name="motion",
     assignments=[
@@ -32,7 +96,7 @@ drag = VariablePortedObject(
 
 drag_magnitude = FunctionalPortedObject(
     name="drag_magnitude",
-    assignments=[("mu", "(coeff*air_density*area)/(2*mass)")]
+    assignments=[("mu", "frac_0(1/2*coeff*rho*area,mass,0)")]
 )
 
 speed = FunctionalPortedObject(
@@ -40,15 +104,15 @@ speed = FunctionalPortedObject(
     assignments=[("s", "sqrt(v_x**2 + v_y**2)")]
 )
 
-proj = CompositePortedObject(
+model = CompositePortedObject(
     name="proj",
     children=[motion, drag, drag_magnitude, speed],
-    input_ports=["coeff", "area", "mass"],
+    input_ports=["C", "A", "m"],
     variable_ports=["pos_x", "pos_y", "vel_x", "vel_y"],
     directed_wires=[
-        ("coeff", "drag_magnitude.coeff"),
-        ("area", "drag_magnitude.area"),
-        ("mass", "drag_magnitude.mass"),
+        ("C", "drag_magnitude.coeff"),
+        ("A", "drag_magnitude.area"),
+        ("m", "drag_magnitude.mass"),
         ("drag_magnitude.mu", "drag.mu"),
         ("motion.v_x", "speed.v_x"),
         ("motion.v_y", "speed.v_y"),
@@ -62,23 +126,27 @@ proj = CompositePortedObject(
     ]
 )
 
-system = System()
-system.add_system_parameter("g", "9.81"),
-system.add_system_parameter("air_density", "1.225")
-system.set_object(proj)
+"""
+Planar motion simulation
+"""
+
[…]
