+++ examples/second_order_ODEs/2-three_body_problem.py
@@ -10,45 +10,40 @@ from psymple.build import (
     System,
 )
 
-S = System(
-    system_parameters=[("G", 1)], #6.67
+"""Common objects"""
+
+vars = VariablePortedObject(
+    name="vars",
+    assignments=[
+        ("v_x", "mu * Del_x"),
+        ("v_y", "mu * Del_y"),
+    ]
 )
 
-# VELOCITY
+dist = FunctionalPortedObject(
+    name="dist",
+    assignments=[
+        ("Del_x", "x_o - x"),
+        ("Del_y", "y_o - y"),
+        ("d", "sqrt((x_o-x)**2 + (y_o-y)**2)"),
+    ]
+)
 
-class force(FunctionalPortedObject):
-    def __init__(self):
-        super().__init__(
-            name=f"force",
-            assignments=[("mu", "G*m/d**3")]
-        )
+force = FunctionalPortedObject(
+    name=f"force",
+    assignments=[("mu", "G*m/d**3")]
+)
 
-class vars(VariablePortedObject):
-    def __init__(self):
-        super().__init__(
-            name="vars",
-            assignments=[
-                ("v_x", "mu * Del_x"),
-                ("v_y", "mu * Del_y"),
-            ]
-        )
-
-class dist(FunctionalPortedObject):
-    def __init__(self):
-        super().__init__(
-            name="dist",
-            assignments=[
-                ("Del_x", "x_o - x"),
-                ("Del_y", "y_o - y"),
-                ("d", "sqrt((x_o-x)**2 + (y_o-y)**2)"),
-            ]
-        )
+"""
+Velocity model
+"""
 
 class velocity(CompositePortedObject):
+
     def __init__(self, id):
         super().__init__(
             name=f"velocity_{id}",
-            children=[vars(), force(), dist()],
+            children=[vars, force, dist],
             input_ports=["x", "y", "x_o", "y_o", "m_o"],
             variable_ports=["v_x", "v_y"],
             directed_wires=[
@@ -68,6 +63,10 @@ class velocity(CompositePortedObject):
             ]
         )
 
+"""
+Trajectory model
+"""
+
 class pos(VariablePortedObject):
     def __init__(self, id):
         super().__init__(
@@ -77,37 +76,54 @@ class pos(VariablePortedObject):
                 ("y", "v_y"),
             ]
         )
-    
-n = 3
 
-bodies = [i for i in range(1,n+1)]
-ints = [f"{i}{j}" for i in range(1,n+1) for j in range(1,n+1) if i != j]
+
+"""
+System data
+"""   
+
+n=3
+
+bodies = [f"{i+1}" for i in range(n)]
+ints = [f"{i+1},{j+1}" for i in range(n) for j in range(n) if i != j]
 coords = ["x", "y"]
 
-system = CompositePortedObject(
+"""
+n-body model
+"""
+
+forces = [velocity(id) for id in ints]
+positions = [pos(id) for id in bodies]
+
+n_body_model = CompositePortedObject(
     name="system",
-    children = [velocity(id) for id in ints] + [pos(i) for i in bodies],
-    variable_ports=[port for i in bodies for port in [f"x_{i}", f"y_{i}", f"v_x_{i}", f"v_y_{i}"]],
+    children = forces + positions,
+    variable_ports=[
+        port 
+        for i in bodies 
+        for port in [f"x_{i}", f"y_{i}", f"v_x_{i}", f"v_y_{i}"]
+    ],
     input_ports=[f"m_{i}" for i in bodies],
[…]
