+++ docs/examples/population_dynamics/predator_prey.md
@@ -1,6 +1,6 @@
 # Predator-prey systems
 
-!!! info "Raw code"
+??? info "Raw code"
 
     The raw code for this example without explanations can be found [here](https://github.com/casasglobal-org/psymple/blob/main/examples/population_dynamics/3-predator_prey.py).
 
