+++ docs/examples/mixing_problems/multiple_tanks.md
@@ -1,6 +1,6 @@
 # Multiple tanks mixing problem
 
-!!! info "Raw code"
+??? info "Raw code"
 
     The raw code for this example without explanations can be found [here](https://github.com/casasglobal-org/psymple/blob/main/examples/mixing_problems/2-multiple_tanks.py).
 
