+++ docs/examples/population_dynamics/logistic_growth.md
@@ -1,6 +1,6 @@
 # The logistic equation
 
-!!! info "Raw code"
+??? info "Raw code"
 
     The raw code for this example without explanations can be found [here](https://github.com/casasglobal-org/psymple/blob/main/examples/population_dynamics/2-logistic_population.py).
 
