### 2025-01-22 casasglobal-org/psymple#59 @GJHSimmons
Thank you for your patience, I've now had time to work through your comments.

This issue comes from LaTeX delimiters inside comments in the code, which are not needed. All the examples have now been updated, and should no longer give this warning when run.

All examples are now documented fully separately. Please see the development version of the documentation [here](https://casasglobal-org.github.io/psymple/dev/examples/).

Also applies to #60 and #61.
