### 2025-01-22 casasglobal-org/psymple#52 @GJHSimmons
Thank you for this. The statement of need, while long, is necessary to track the previous contributions in this area.

The obvious candidate for removal is the non-standard "Description" section. The authors believe it makes the manuscript read better, but we are happy to move this section to the documentation website if you prefer, and reference it from the paper.

Please advise further.
