### 2025-01-22 casasglobal-org/psymple#62 @GJHSimmons
I've pushed an update to the main branch which supresses this warning.

The remaining user warnings from running this example, for example:

```
UserWarning: The parameter 'g' has been specified as a system parameter. A port for 'g' will not be created in 'v_gravity'
```

are expected behaviour.

Also applies to #63.
