+++ tox.ini
@@ -1,5 +1,9 @@
 [tox]
-envlist = py{310}
+envlist = 
+    py{310}
+    py{311}
+    py{312}
+    py{313}
 minversion = 4.0
 isolated_build = true
 
@@ -12,3 +16,6 @@ commands = pytest tests
 [gh]
 python =
     3.10 = py310
+    3.11 = py311
+    3.12 = py312
+    3.13 = py313
