+++ mkdocs.yml
@@ -125,4 +125,5 @@ nav:
     - Get involved: development/development.md
     - Development team: development/development_team.md
     - Community guidelines: development/community_guidelines.md
+    - Testing psymple: development/testing_psymple.md
   
