### 2025-01-23 casasglobal-org/psymple#63 @pitmonticone
Thanks, I consider the issue resolved.
