### 2025-01-23 casasglobal-org/psymple#59 @pitmonticone
Thanks, I consider the issue as resolved.
