### 2025-01-23 casasglobal-org/psymple#60 @pitmonticone
Thanks, I consider the issue as resolved.
