### 2025-01-23 casasglobal-org/psymple#61 @pitmonticone
Thanks, I consider the issue as resolved.
