### 2025-01-23 #7364 @SunnyXu
@GJHSimmons
Hi George,
Thanks for your great improvement of the documentation. I think you have addressed all the issues I raised related to now. Therefore, I have updated my review checklist.
Regarding your word counts in the manuscript (https://github.com/casasglobal-org/psymple/issues/52), if the editor does not mind about this, I am fine with it as well.
@prashjha
Hi Prashant,
Please let me know if I need to do anything else. Thanks again for considering me as a reviewer. Please do not hesitate to contact me again if I can help with any review in the future.
Jin
