+++ mkdocs.yml
@@ -67,7 +67,6 @@ plugins:
 extra:
   version:
     provider: mike
-    default: latest
 
 nav:
   - Home: index.md
