+++ mkdocs.yml
@@ -67,6 +67,7 @@ plugins:
 extra:
   version:
     provider: mike
+    default: latest
 
 nav:
   - Home: index.md
