+++ mkdocs.yml
@@ -68,6 +68,7 @@ extra:
   version:
     provider: mike
     default: latest
+    alias: true
 
 nav:
   - Home: index.md
