+++ .github/workflows/docs.yml
@@ -2,6 +2,7 @@ name: build docs
 on:
   release:
     types: [created]
+  workflow_dispatch:
 permissions:
   contents: write
 jobs:
