### 2025-01-23 33b5ef1
Delete latest directory
