### 2025-01-23 f99f66c
docs latex error update
