+++ README.md
@@ -20,10 +20,10 @@ The latest version of `psymple` documentation can be found [here](https://casasg
 
 For most new users, going through the following resources in the suggested order will provide a good introduction to `psymple`. 
 
-1. To familiarise yourself with the need, purpose and design of `psymple`, the best place to start is [here](https://casasglobal-org.github.io/psymple/dev/overview/introduction/).
-2. To start modelling with `psymple`, the best place to find out about capturing functions and differential equations is [here](https://casasglobal-org.github.io/psymple/dev/components/functional_ported_objects/).
-3. To start simulating models in `psymple`, start reading about systems [here](https://casasglobal-org.github.io/psymple/dev/user_guide/system/).
-4. A detailed set of examples demonstrating the process of building and simulating models can be found [here](https://casasglobal-org.github.io/psymple/dev/examples/).
-5. If you're interested in the mathematical basis of `psymple`, the best place to start is [here](https://casasglobal-org.github.io/psymple/dev/mathematics/ported_objects/).
-6. Information on how to get help, support, or get involved with developing `psymple` can be found [here](https://casasglobal-org.github.io/psymple/dev/development/development/).
+1. To familiarise yourself with the need, purpose and design of `psymple`, the best place to start is [here](https://casasglobal-org.github.io/psymple/latest/overview/introduction/).
+2. To start modelling with `psymple`, the best place to find out about capturing functions and differential equations is [here](https://casasglobal-org.github.io/psymple/latest/components/functional_ported_objects/).
+3. To start simulating models in `psymple`, start reading about systems [here](https://casasglobal-org.github.io/psymple/latest/user_guide/system/).
+4. A detailed set of examples demonstrating the process of building and simulating models can be found [here](https://casasglobal-org.github.io/psymple/latest/examples/).
+5. If you're interested in the mathematical basis of `psymple`, the best place to start is [here](https://casasglobal-org.github.io/psymple/latest/mathematics/ported_objects/).
+6. Information on how to get help, support, or get involved with developing `psymple` can be found [here](https://casasglobal-org.github.io/psymple/latest/development/development/).
 
