+++ examples/README.md
@@ -1,0 +1,3 @@
+# Examples
+
+The examples contained in this directory consist of raw code with few comments or explanations. Documented versions of these examples can be found on the [examples page](https://casasglobal-org.github.io/psymple/latest/examples/) of the documentation.
