### 2025-01-24 casasglobal-org/psymple#66 @pitmonticone
[REVIEW]: Feedback on the paper
Dear authors,

Overall, I believe that the paper is well-written and satisfies most of the criteria in the JOSS checklist:

- [x] **Summary**: No revision required.
- [x]  **Statement of need**: No revision required.
- [x] **State of the field**: I believe this could be improved by adding a "Related packages" section. This would give readers a more comprehensive understanding of the existing tools and how `psymple` distinguishes itself.
- [x] **Quality of writing**: No revision required.
- [x]  **References**: No revision required.

REVIEW ISSUE: https://github.com/openjournals/joss-reviews/issues/7364
