### 2025-03-06 #7364 @GJHSimmons
@prashjha I'm waiting for advice on how we should reduce the length. For example, we could move the "Description" section to the documentation, see [here](https://github.com/casasglobal-org/psymple/issues/52) or try to cut words here and there. The former would be easier for us.

I can do this together with updating the paper with @pitmonticone's final comments. Please advise.
