### 2025-03-14 698755a
allow outputs from child variables
