### 2025-03-23 #7364 @prashjha
> [@prashjha](https://github.com/prashjha) I'm waiting for advice on how/how much we should reduce the length. For example, we could move the "Description" section to the documentation, see [here](https://github.com/casasglobal-org/psymple/issues/52), or try to cut words here and there. The former would be easier for us.
>
> I can do this together with updating the paper with [@pitmonticone](https://github.com/pitmonticone)'s final comments. Please advise.

Hi @GJHSimmons, see what you can do to reduce the length of the paper without spending too much time or compromising the original version. If you think you can not reduce the length, just let me know and it should be okay.

Also, there are a few issues from @pitmonticone that need to be addressed.

Once the draft is updated and the issues are addressed, let me and @pitmonticone so we can move to the next stage.
