+++ paper.md
@@ -60,19 +60,17 @@ While PBDM is a highly-developed framework, it shares the same barriers to wides
 
 Models in `psymple` are built from arbitrary combinations of modular mechanistic, dynamic components and correlative, functional components which naturally interact with each other. This structure allows for the systematic implementation of modelling frameworks such as PBDM, and, more widely, those capturing biological, economic and social systems, for which it is not feasible to capture the laws of interaction purely mechanistically. Examples include agroecological, bioeconomic and Earth systems modelling, and the development of `psymple` is a necessary first step in the development and release of accessible and impactful tools in these areas. 
 
-# Description
+# Description and related software
 
-The workings of `psymple` are based on the ideas of the category-theoretic dynamic systems modelling package `AlgebraicJulia.jl` [@l-b-p-f:2024], in which modular ‘ported’ objects, called *resource sharers*, containing ordinary differential equations, expose internal variable states at different ports. Ports are linked by wires which represent the sharing of state variables across the dynamics equations of the system to create complex interacting dynamical systems. Mathematically, resource sharers form algebras over the operad of undirected wiring diagrams, see [@l-b-p-f:2022]. 
+The workings of `psymple` are based on the ideas of the dynamic systems modelling package `AlgebraicJulia.jl` [@l-b-p-f:2024], in which modular objects, called *resource sharers*, containing ordinary differential equations, are linked by wires which represent the sharing of state variables across the dynamics equations of the system. A compilation process based in mathematical category theory [@l-b-p-f:2022] maps this diagram of objects and wires into a complex interacting dynamical system. 
 
-In `psymple`, these ideas are extended to realise functions as ported objects, called functional ported objects (FPOs), alongside resource sharers, which in `psymple` are called variable ported objects (VPOs). FPOs contain systems of multivariate functions, whose values are exposed at output ports, and whose arguments can read information in from other FPO outputs or from VPO variable ports. The structure of input ports is also introduced to VPOs by allowing functions defining differential equations to read information in the same way as FPOs.
+In `psymple`, these ideas are extended to allow for a second type of object which compute multivariate functions. Wires between these objects formally represent partial functional substitution, and together with dynamic resource sharers enable arbitrary hybrid models of complex systems to be created, compiled, and simulated. Underneath, the mathematical conversion from formal system diagrams of differential equations and functions to simulatable systems is handled by the Python symbolic mathematics package `sympy` [@meur:2017].
 
-FPOs and VPOs can be arbitrarily combined into composite ported objects (CPOs) through the concepts of directed wires, which capture partial substitution of functional information, and variable wires, which implement the functionality of resource sharing. CPOs can themselves read and expose information from input, output and variable ports to create fully modular and arbitrarily complex hybrid systems of both functional and dynamic components whose nested hierarchy reflects the model structure. 
+Additionally, `psymple` enables arbitrary nesting of these base objects inside composite objects whose hierarchy represents modelled structure. With these structures, `psymple` satisfies a specification shared by "next-generation" dynamical systems modelling frameworks, see [@baez:2023], including being *faceted*, where models can be considered one piece at a time; *modular*, where components naturally compose together; and *functorial*, where the data describing the model (its syntax) is systematically and reliably transformed into system behaviour (its semantics). These ideas allow for legible modelling of highly complex, specialised systems, and drive clear, adaptable and accessible modelling practices in line with those promoted by the [Open Modeling Foundation](https://www.openmodelingfoundation.org/). 
 
[…]
