+++ paper.bib
@@ -265,6 +265,36 @@ doi = {10.1079/9781800621459.0000}
   url={https://api.semanticscholar.org/CorpusID:135291467}
 }
 
+@article{holz+:2014,
+  author = {Holzworth, D.P. and Huth, N.I. and deVoil, P.G. and Zurcher, E.J. and Herrmann, N.I. and McLean, G. and Chenu, K. and others},
+  title = {APSIM -- Evolution towards a New Generation of Agricultural Systems Simulation},
+  journal = {Environmental Modelling \& Software},
+  volume = {62},
+  pages = {327--350},
+  year = {2014},
+  doi = {10.1016/j.envsoft.2014.07.009}
+}
+
+@inproceedings{spor+:2009,
+author = {Sporleder, M. and Chavez, D. and Gonzales, J. and Juarez, H. and Simon, R. and Kroschel, J.},
+year = {2009},
+month = {11},
+pages = {},
+title = {ILCYM-Insect life cycle modeling: software for developing temperature-based insect phenology models with applications for regional and global pest risk assessments and mapping}
+}
+
+@Article{jone+:2003,
+  title = {The DSSAT Cropping System Model},
+  author = {Jones, J.W. and Hoogenboom, G. and Porter, C.H.
+    and Boote, K.J. and Batchelor, W.D. and Hunt, L.A.
+    and Wilkens, P.W. and Singh, U. and Gijsman, A.J. and Ritchie, J.T.},
+  journal = {European Journal of Agronomy},
+  volume = {18},
+  pages = {235-265},
+  year = {2003},
+  number = {3-4},
+}
+
 @book{dw-go:1978,
   author={de Wit, C. and Goudrian, J.},
   title={Simulation of ecological processes},
