+++ paper.md
@@ -68,7 +68,7 @@ In `psymple`, these ideas are extended to allow for a second type of object whic
 
 Additionally, `psymple` enables arbitrary nesting of these base objects inside composite objects whose hierarchy represents modelled structure. With these structures, `psymple` satisfies a specification shared by "next-generation" dynamical systems modelling frameworks, see [@baez:2023], including being *faceted*, where models can be considered one piece at a time; *modular*, where components naturally compose together; and *functorial*, where the data describing the model (its syntax) is systematically and reliably transformed into system behaviour (its semantics). These ideas allow for legible modelling of highly complex, specialised systems, and drive clear, adaptable and accessible modelling practices in line with those promoted by the [Open Modeling Foundation](https://www.openmodelingfoundation.org/). 
 
-Many software solutions exist to capture climate-driven behaviour within parts of agricultural ecosystems. For example, sophisticated climate-sensitive crop models can be produced using APSIM [@holz+:2014] or DSSAT [@jone+:2003]. For insects, ILCYM [@spor+:2009] enables the creation of climate-driven pest insect phenology models for distribution and risk assessments. The goal `psymple` and its framework is not to stand apart from this existing software, but instead enable the creation of software which can link between existing tools and enable such assessments across ecosystems, including the plant, herbivore and natural enemy trophic levels and their interactions.
+Many software solutions exist to capture climate-driven behaviour within parts of agricultural ecosystems. For example, sophisticated climate-sensitive crop models can be produced using `APSIM` [@holz:2014] or `DSSAT` [@jone:2003]. For insects, `ILCYM` [@spor:2009] enables the creation of climate-driven pest insect phenology models for distribution and risk assessments. The goal `psymple` and its framework is not to stand apart from this existing software, but instead enable the creation of software which can link between existing tools and enable such assessments across ecosystems, including the plant, herbivore and natural enemy trophic levels and their interactions.
 
 # Related research
 
