+++ paper.bib
@@ -265,7 +265,7 @@ doi = {10.1079/9781800621459.0000}
   url={https://api.semanticscholar.org/CorpusID:135291467}
 }
 
-@article{holz+:2014,
+@article{holz:2014,
   author = {Holzworth, D.P. and Huth, N.I. and deVoil, P.G. and Zurcher, E.J. and Herrmann, N.I. and McLean, G. and Chenu, K. and others},
   title = {APSIM -- Evolution towards a New Generation of Agricultural Systems Simulation},
   journal = {Environmental Modelling \& Software},
@@ -275,7 +275,7 @@ doi = {10.1079/9781800621459.0000}
   doi = {10.1016/j.envsoft.2014.07.009}
 }
 
-@inproceedings{spor+:2009,
+@inproceedings{spor:2009,
 author = {Sporleder, M. and Chavez, D. and Gonzales, J. and Juarez, H. and Simon, R. and Kroschel, J.},
 year = {2009},
 month = {11},
@@ -283,7 +283,7 @@ pages = {},
 title = {ILCYM-Insect life cycle modeling: software for developing temperature-based insect phenology models with applications for regional and global pest risk assessments and mapping}
 }
 
-@Article{jone+:2003,
+@Article{jone:2003,
   title = {The DSSAT Cropping System Model},
   author = {Jones, J.W. and Hoogenboom, G. and Porter, C.H.
     and Boote, K.J. and Batchelor, W.D. and Hunt, L.A.
