+++ paper.md
@@ -64,11 +64,11 @@ Models in `psymple` are built from arbitrary combinations of modular mechanistic
 
 The workings of `psymple` are based on the ideas of the dynamic systems modelling package `AlgebraicJulia.jl` [@l-b-p-f:2024], in which modular objects, called *resource sharers*, containing ordinary differential equations, are linked by wires which represent the sharing of state variables across the dynamics equations of the system. A compilation process based in mathematical category theory [@l-b-p-f:2022] maps this diagram of objects and wires into a complex interacting dynamical system. 
 
-In `psymple`, these ideas are extended to allow for a second type of object which compute multivariate functions. Wires between these objects formally represent partial functional substitution, and together with dynamic resource sharers enable arbitrary hybrid models of complex systems to be created, compiled, and simulated. Underneath, the mathematical conversion from formal system diagrams of differential equations and functions to simulatable systems is handled by the Python symbolic mathematics package `sympy` [@meur:2017].
+In `psymple`, these ideas are extended to allow for a second type of object which computes multivariate functions. Wires between these objects formally represent partial functional substitution, and together with dynamic resource sharers enable arbitrary hybrid models of complex systems to be created, compiled, and simulated. Underneath, the mathematical conversion from formal system diagrams of differential equations and functions to simulatable systems is handled by the Python symbolic mathematics package `sympy` [@meur:2017].
 
 Additionally, `psymple` enables arbitrary nesting of these base objects inside composite objects whose hierarchy represents modelled structure. With these structures, `psymple` satisfies a specification shared by "next-generation" dynamical systems modelling frameworks, see [@baez:2023], including being *faceted*, where models can be considered one piece at a time; *modular*, where components naturally compose together; and *functorial*, where the data describing the model (its syntax) is systematically and reliably transformed into system behaviour (its semantics). These ideas allow for legible modelling of highly complex, specialised systems, and drive clear, adaptable and accessible modelling practices in line with those promoted by the [Open Modeling Foundation](https://www.openmodelingfoundation.org/). 
 
-Many software solutions exist to capture climate-driven behaviour within parts of agricultural ecosystems. For example, sophisticated climate-sensitive crop models can be produced using `APSIM` [@holz:2014] or `DSSAT` [@jone:2003]. For insects, `ILCYM` [@spor:2009] enables the creation of climate-driven pest insect phenology models for distribution and risk assessments. The goal `psymple` and its framework is not to stand apart from this existing software, but instead enable the creation of software which can link between existing tools and enable such assessments across ecosystems, including the plant, herbivore and natural enemy trophic levels and their interactions.
+Many software solutions exist to capture climate-driven behaviour within parts of agricultural ecosystems. For example, sophisticated climate-sensitive crop models can be produced using `APSIM` [@holz:2014] or `DSSAT` [@jone:2003]. For insects, `ILCYM` [@spor:2009] enables the creation of climate-driven pest insect phenology models for distribution and risk assessments. The goal of `psymple` and its framework is not to stand apart from this existing software, but instead enable the creation of software which can link between existing tools and enable such assessments across ecosystems, including the plant, herbivore and natural enemy trophic levels and their interactions.
 
 # Related research
 
