### 2025-03-27 casasglobal-org/psymple#66 @GJHSimmons
@pitmonticone I have combined a stripped-down description section with a brief survey of related existing tools. Please let me know if this is in the format expected.
