### 2025-03-27 #7364 @GJHSimmons
@prashjha I have completed the updates, combining a stripped-down description section (where information is already in the documentation) with the survey of existing software requested by @pitmonticone. The net result is about the same length of manuscript. We couldn't identify any other places to effectively reduce the length. Please let me know if that is okay.
