### 2025-03-27 casasglobal-org/psymple#66 @pitmonticone
It looks good to me.
