### 2025-04-11 bc0f50b
outputs from child variable ports
