### 2025-04-14 #7364 @prashjha
> [@prashjha](https://github.com/prashjha) I have completed the updates, combining a stripped-down description section (where information is already in the documentation) with the survey of existing software requested by [@pitmonticone](https://github.com/pitmonticone). The net result is about the same length of manuscript. We couldn't identify any other places to effectively reduce the length. Please let me know if that is okay.

I am alright with the length of the paper. However, the final decision is made by EiC, who may have a different view. We will see how it goes for now.
