### 2025-04-14 #7364 @prashjha
Hi @GJHSimmons, I am reading the draft and will let you know if I have any suggestions.

Meanwhile, could you please archive (if not done already) the release using zenodo and provide the archive reference so that I can associate it with your JOSS submission? Also, please ensure that the zenodo archive's title matches the title of this JOSS article.

If you have updated a version of your code, let me know, and I can update it here.
