+++ psymple/build/ported_objects.py
@@ -95,7 +95,7 @@ class PortedObjectData(dict):
         
     def _check_object_data(self, data):
         # Checks that the format of object_data is parsable
-        ALLOWED_KEYS = {"assignments", "input_ports", "variable_ports", "output_ports", "variable_wires", "directed_wires", "children", "create_input_ports"}
+        ALLOWED_KEYS = {"assignments", "input_ports", "variable_ports", "output_ports", "variable_wires", "directed_wires", "children", "create_input_ports", "create_output_ports"}
         if extra_keys := data.keys() - ALLOWED_KEYS:
             raise KeyError(f"Object data {data} cannot include keys {extra_keys}")
         
@@ -160,6 +160,7 @@ class VariablePortedObject(PortedObjectWithAssignments):
         input_ports: list[InputPort | dict | tuple | str] = [],
         variable_ports: list[InputPort | dict | str] = [],
         assignments: list[DifferentialAssignment | dict | tuple] = [],
+        create_output_ports: bool = True,
         create_input_ports: bool = True,
         parsing_locals: dict = {},
     ):
@@ -175,6 +176,8 @@ class VariablePortedObject(PortedObjectWithAssignments):
                 See [add_variable_ports][psymple.build.abstract.PortedObject.add_variable_ports].
             assignments: list of differential assignments (ODEs). See 
                 [add_variable_assignments][psymple.build.VariablePortedObject.add_variable_assignments]
+            create_output_ports: if `True`, automatically create an output port exposing each exposed variable
+                to be read by directed wires. See Notes for more information.
             create_input_ports: if `True`, automatically expose all parameters as input ports. See Notes for more
                 information.
             parsing_locals: a dictionary mapping strings to `sympy` objects.
@@ -183,6 +186,9 @@ class VariablePortedObject(PortedObjectWithAssignments):
             - By default, each variable (dependent variable of each ODE) is automatically exposed as a
                 variable port. Alternatively, chosen variables can be exposed by specifying
                 them in the list variable_ports.
+
+            - By default, each exposed variable will be mirrored in the creation of an output port with
+                the same name. This output port reads the value of the variable.
 
             - Parameters listed in input_ports are exposed and can be used in ODE expressions.
 
@@ -203,14 +209,17 @@ class VariablePortedObject(PortedObjectWithAssignments):
         self.add_variable_assignments(
             *assignments,
             create_variable_ports=create_variable_ports,
+            create_output_ports=create_output_ports,
             create_input_ports=create_input_ports,
         )
         self.create_input_ports = create_input_ports
+        self.create_output_ports = create_output_ports
 
     def add_variable_assignments(
         self,
         *assignments: DifferentialAssignment | dict | tuple,
         create_variable_ports: bool = True,
+        create_output_ports: bool = True,
         create_input_ports: bool = True,
     ):
         """
@@ -229,6 +238,8 @@ class VariablePortedObject(PortedObjectWithAssignments):
 
             create_variable_ports: if `True`, variable ports exposing the variable of each assignment
                 will automatically be created.
+            create_output_ports: if `True`, output ports mirroring each variable port will
+                automatically be created.
             create_input_ports: if `True`, input ports for each free symbol in the expression of each
                 assignment will automatically be created.
 
@@ -249,6 +260,10 @@ class VariablePortedObject(PortedObjectWithAssignments):
                 self.add_variable_ports(variable_name)
             elif variable_name not in self.variable_ports:
                 self.internals[variable_name] = assignment.variable
+
+        if create_output_ports:
+            self.add_output_ports(*self.variable_ports.keys())
+                
         if create_input_ports:
             # Create input ports for all symbols that are not variables (exposed or internal) or already
             # specified as input ports.
[…]
