### 2025-04-17 #7364 @GJHSimmons
@prashjha

I have archived the package as [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15237709.svg)](https://doi.org/10.5281/zenodo.15237709).

The title, authors, affiliations and license have been updated to match the submitted article.

The archived release is version 1.0.0 encompassing all changes from the review.
