### 2025-04-29 #7364 @GJHSimmons
@prashjha

I hope you're well. A small favour to ask: a collaborator previously thanked in acknowledgements helped to directly steer some changes in the most recent release. We would like to now credit them as a contributing author. I appreciate this is late in the process, but if it is okay with you I will update the draft with their details, as well as the [development team page](https://casasglobal-org.github.io/psymple/latest/development/development_team/) confirming the contribution, and the zenodo release.

I would also at the same time like to update the draft to fix some grammatical errors and cite a new reference. I'll prepare a list of the changes made so they don't interfere with your review.
