+++ paper.bib
@@ -8,6 +8,20 @@ isbn = {9780128096338},
 doi = {10.1016/B978-0-12-809633-8.02390-6}
 }
 
+@article{flat2011,
+  author = {Flato, G. M.},
+  title = {Earth system models: an overview},
+  journal = {WIREs Climate Change},
+  volume = {2},
+  number = {6},
+  pages = {783-800},
+  doi = {https://doi.org/10.1002/wcc.148},
+  url = {https://wires.onlinelibrary.wiley.com/doi/abs/10.1002/wcc.148},
+  eprint = {https://wires.onlinelibrary.wiley.com/doi/pdf/10.1002/wcc.148},
+  abstract = {Abstract Earth System models (ESMs) are global climate models with the added capability to explicitly represent biogeochemical processes that interact with the physical climate and so alter its response to forcing such as that associated with human-caused emissions of greenhouse gases. Representing the global carbon cycle allows for feedbacks between the physical climate and the biological and chemical processes in the ocean and on land that take up some of the emitted carbon dioxide and so act to reduce warming. The sulfur cycle is also important in that both natural and human emissions of sulfur contribute to the production of sulfate aerosols which reflect incoming solar radiation (a direct cooling effect) and alter cloud properties (an indirect cooling effect). Other components such as ozone are also being incorporated into some ESMs. Evaluating the physical component of an ESM is becoming increasingly comprehensive and sophisticated, but the evaluation of the biogeochemical components suffer somewhat from a lack of comprehensive global-scale observational data. Nevertheless, such models provide valuable insight into climate variability and change, and the role of human activities and possible mitigation actions on future climate change. Internationally coordinated experiments are increasingly important in providing a multimodel ensemble of climate simulations, thereby taking advantage of some ‘cancellation of errors’ and allowing better quantification of uncertainty. WIREs Clim Change 2011, 2:783–800. doi: 10.1002/wcc.148 This article is categorized under: Climate Models and Modeling > Earth System Models},
+  year = {2011}
+}
+
 @article{el-le:2009,
    author = "Elith, Jane and Leathwick, John R.",
    title = "Species Distribution Models: Ecological Explanation and Prediction Across Space and Time", 
@@ -84,6 +98,17 @@ pages = {},
 title = {Challenging the status quo in invasive species assessment using mechanistic physiologically based demographic modeling},
 journal = {Environment, Development and Sustainability},
 doi = {10.1007/s10668-023-03698-9}
+}
+
+@article{guti+:2025b,
+author = {Gutierrez, A.P. and Ponti, L. and Neteler, M. and Stocchino, F. and Cure, J.R. and Kenmore, P.E. and Simmons, G.},
+year = {2025},
+month = {04},
+pages = {},
+title = {Geospatial physiologically based demographic modeling and analysis of thirteen invasive species in Africa: why the biology matters},
+volume = {37},
+journal = {Environmental Sciences Europe},
+doi = {10.1186/s12302-025-01084-y}
 }
 
 @article{g-y-n-e:1999,
@@ -192,8 +217,24 @@ journal = {Journal of Pest Science},
 doi = {10.1007/s10340-023-01610-y}
 }
 
+@article{gu-re:2005,
+  title = {The bioeconomics of tritrophic systems: applications to invasive species},
+  journal = {Ecological Economics},
+  volume = {52},
+  number = {3},
+  pages = {383-396},
+  year = {2005},
+  note = {Integrating Ecology and Economics in Control Bioinvasions},
+  issn = {0921-8009},
+  doi = {https://doi.org/10.1016/j.ecolecon.2004.05.007},
[…]
