+++ paper.md
@@ -19,6 +19,9 @@ authors:
   - name: Chiara Facciolà
     orcid: 0000-0001-8359-9300
     affiliation: 1
+  - name: Timothy Hosgood
+    orcid: 0000-0003-4580-0741
+    affiliation: 5
 
 affiliations:
  - name: IDEMS International CIC, United Kingdom
@@ -32,6 +35,8 @@ affiliations:
  - name:
     Division of Ecosystem Science, College of Natural Resources, University of California, Berkeley, USA
    index: 4
+ - name:
+    Topos Institute, Berkeley, USA
 date: 13 September 2024
 bibliography: paper.bib
 ---
@@ -50,15 +55,16 @@ The development of `psymple` emerged from the complex system modelling requireme
 
 An alternative approach is mechanistic modelling, which instead of observational data use physiological data to capture the underlying mechanisms which drive species distribution, such as energy balance, population dynamics or response to climate [@ke-po:2009]. In contrast to correlative approaches, mechanistic SDMs decouple the physiology of a species from their native geography or climate, and allow SDMs in new geographic or climatic regimes to be created in the absence of observational data [@john:2019]. 
 
-An example mechanistic framework is physiologically-based demographic modelling (PBDM), which creates holistic ecosystem models based on the weather-driven biology of component species, allowing for predictive phenology, age- or mass-structured population dynamics, and distribution assessments, see [@gu-po:2022a; @gu-po:2022b] for an overview and further references. With this approach, PBDM can account for tritrophic ecosystem interactions [@g-y-n-e:1999], or model the effects of climate change [@guti:2023].
+An example mechanistic framework is physiologically-based demographic modelling (PBDM), which creates holistic ecosystem models based on the weather-driven biology of component species, allowing for predictive phenology, age- or mass-structured population dynamics, and distribution assessments, see [@gu-po:2022a; @gu-po:2022b; @guti+:2025a] for an overview and further references. With this approach, PBDM can account for tritrophic ecosystem interactions [@g-y-n-e:1999], or model the effects of climate change [@guti:2023].
 
-The schools of thought around correlative SDMs and mechanistic frameworks such as PBDM are seen as largely disjoint [@dorm:2012], but their development can be traced back to early common roots [@fi-ni:1970; @guti:1974; @dw-go:1978]. A component of the PBDM framework is the use of physiological data to parametrise "biodemographic" functions capturing biophysical or biochemical mechanisms, such as the development, mortality and fecundity rates of a species in response to environmental variables [@po-gu:2023].
+While correlative SDMs and mechanistic frameworks such as PBDM are today often regarded as conceptually distinct and largely unintegrated [@dorm:2012], their development can be traced back to early common roots [@fi-ni:1970; @guti:1974; @dw-go:1978]. A component of the PBDM framework is the use of physiological data to parametrise "biodemographic" functions capturing biophysical or biochemical mechanisms, such as the development, mortality and fecundity rates of a species in response to environmental variables [@po-gu:2023].
 
 The use of biodemographic functions in PBDM combines the considerable holistic advantages of mechanistic SDMs, while retaining the comparable ease of parametrisation as status-quo correlative models. More widely, there has been growing interest and development of ecological models explicitly composed of both correlative and mechanistic components [@buck:2010; @to-va:2023], combining the benefits of both areas. In the wider context of complex systems modelling, the approach of building composite models out of different techniques is called hybrid, or spectrum, modelling.  
 
[…]
