+++ paper.bib
@@ -8,7 +8,7 @@ isbn = {9780128096338},
 doi = {10.1016/B978-0-12-809633-8.02390-6}
 }
 
-@article{flat2011,
+@article{flat:2011,
   author = {Flato, G. M.},
   title = {Earth system models: an overview},
   journal = {WIREs Climate Change},
@@ -336,7 +336,7 @@ title = {ILCYM-Insect life cycle modeling: software for developing temperature-b
   number = {3-4},
 }
 
-@article{guti+:2025a,
+@article{guti:2025a,
   author = {Gutierrez, A.P. and Ponti, L. and Levi-Mourao, A. and Pons, X. and Cure, J.R. and Neteler, M. and Simmons, G.},
   title = {Stabilizing adaptation in an invasive species: {A}lfalfa weevil as a case study},
   journal = {Agricultural and Forest Entomology},
