+++ paper.bib
@@ -100,7 +100,7 @@ journal = {Environment, Development and Sustainability},
 doi = {10.1007/s10668-023-03698-9}
 }
 
-@article{guti+:2025b,
+@article{guti:2025b,
 author = {Gutierrez, A.P. and Ponti, L. and Neteler, M. and Stocchino, F. and Cure, J.R. and Kenmore, P.E. and Simmons, G.},
 year = {2025},
 month = {04},
@@ -340,9 +340,7 @@ title = {ILCYM-Insect life cycle modeling: software for developing temperature-b
   author = {Gutierrez, A.P. and Ponti, L. and Levi-Mourao, A. and Pons, X. and Cure, J.R. and Neteler, M. and Simmons, G.},
   title = {Stabilizing adaptation in an invasive species: {A}lfalfa weevil as a case study},
   journal = {Agricultural and Forest Entomology},
-  volume = {n/a},
-  number = {n/a},
-  pages = {},
+  year = {2025},
   keywords = {geographic distribution, invasive species, pest risk analysis, physiologically based demographic models, population dynamics, relative abundance, stabilizing selection, strains},
   doi = {https://doi.org/10.1111/afe.12686},
   url = {https://resjournals.onlinelibrary.wiley.com/doi/abs/10.1111/afe.12686},
