+++ paper.md
@@ -64,7 +64,7 @@ The use of biodemographic functions in PBDM combines the considerable holistic a
 
 While PBDM is a highly-developed framework, it shares the same barriers to widespread introduction as general hybrid complex system models. These barriers include the lack of available modelling frameworks [@bu-c-j:2018], the lack of flexibility in existing models [@buck:2010], or the lack of modelling platforms to implement existing ideas [@guti:2025b; @po-gu:2023]. The package `psymple` is a general platform designed to facilitate the creation of hybrid complex systems models and modelling frameworks. 
 
-Models in `psymple` are built from arbitrary combinations of modular mechanistic, dynamic components and correlative, functional components which naturally interact with each other. This structure allows for the systematic implementation of modelling frameworks such as PBDM, and, more widely, those capturing biological, economic and social systems, for which it is not feasible to capture the laws of interaction purely mechanistically. Examples include agroecological and bioeconomic models (see [@guti:2020; @gu-re:2005; r-g-s-z:1998]), and Earth system models capturing the biogeochemical interactions between the ecosystems, humanity, and climate [@flat:2011]. The development of of `psymple` is a necessary first step in the development and release of accessible and impactful tools in these areas. 
+Models in `psymple` are built from arbitrary combinations of modular mechanistic, dynamic components and correlative, functional components which naturally interact with each other. This structure allows for the systematic implementation of modelling frameworks such as PBDM, and, more widely, those capturing biological, economic and social systems, for which it is not feasible to capture the laws of interaction purely mechanistically. Examples include agroecological and bioeconomic models (see [@guti:2020; @gu-re:2005; @r-g-s-z:1998]), and Earth system models capturing the biogeochemical interactions between the ecosystems, humanity, and climate [@flat:2011]. The development of of `psymple` is a necessary first step in the development and release of accessible and impactful tools in these areas. 
 
 
 # Description and related software
@@ -83,7 +83,7 @@ The development of `psymple` emerged as part of a collaborative effort between [
 
 The specification of `psymple` was created from the "collaborative modelling" approach headed-up by the [Topos Institute](https://topos.institute/), which uses the mathematical field of applied category theory to capture the mechanisms required to create modelling frameworks collaboratively, accounting for diversity in technical language, data availability and subject expertise. This specification, coupled with the generality and flexibility of `psymple`, enables frameworks beyond PBDM to be developed or implemented, and pushed towards new research and subsequent impacts.
 
-For example, research is currently active in bioeconomic modelling, which extends the supply-demand and mass-structured mechanisms of PBDM to understand the economic interaction that humans have with agroecological systems, see [@guti:2020; @r-g-s-z:1998]. More widely, ideas are being conceived around applications in Earth systems modelling, multi-layer modelling incorporating feedback loops to modelling techniques such as [collaborative agent-based modelling](https://johncarlosbaez.wordpress.com/2023/08/17/agent-based-models-part-2/), and data management techniques such as source control and versioning across collaborative practices.
+For example, research is currently active in bioeconomic modelling, which extends the supply-demand and mass-structured mechanisms of PBDM to understand the economic interaction that humans have with agroecological systems, see [@guti:2020; @r-g-s-z:1998]. More widely, ideas are being conceived around applications in Earth system modelling [@flat:2011], multi-layer modelling incorporating feedback loops to modelling techniques such as [collaborative agent-based modelling](https://johncarlosbaez.wordpress.com/2023/08/17/agent-based-models-part-2/), and data management techniques such as source control and versioning across collaborative practices.
 
 # Acknowledgements
 
