+++ paper.bib
@@ -1,5 +1,5 @@
 @inbook{el-fr:2017,
-author = {Elith, Jane and Franklin, Janet},
+author = {Elith, J. and Franklin, J.},
 year = {2017},
 month = {01},
 pages = {},
@@ -23,7 +23,7 @@ doi = {10.1016/B978-0-12-809633-8.02390-6}
 }
 
 @article{el-le:2009,
-   author = "Elith, Jane and Leathwick, John R.",
+   author = "Elith, J. and Leathwick, J.R.",
    title = "Species Distribution Models: Ecological Explanation and Prediction Across Space and Time", 
    journal= "Annual Review of Ecology, Evolution, and Systematics",
    year = "2009",
@@ -34,7 +34,6 @@ doi = {10.1016/B978-0-12-809633-8.02390-6}
    url = "https://www.annualreviews.org/content/journals/10.1146/annurev.ecolsys.110308.120159",
    publisher = "Annual Reviews",
    issn = "1545-2069",
-   type = "Journal Article",
    keywords = "climate change",
    keywords = "niche",
    keywords = "spatial",
@@ -45,7 +44,7 @@ doi = {10.1016/B978-0-12-809633-8.02390-6}
   }
 
 @article{ke-po:2009,
-author = {Kearney, Michael and Porter, Warren},
+author = {Kearney, M. and Porter, W.},
 title = {Mechanistic niche modelling: combining physiological and spatial data to predict species’ ranges},
 journal = {Ecology Letters},
 volume = {12},
@@ -76,7 +75,7 @@ eprint = {https://royalsocietypublishing.org/doi/pdf/10.1098/rspb.2019.1916}
 }
 
 @article{to-va:2023,
-author = {Tourinho, Luara and Vale, Mariana M.},
+author = {Tourinho, L. and Vale, M.M.},
 title = {Choosing among correlative, mechanistic, and hybrid models of species’ niche and distribution},
 journal = {Integrative Zoology},
 volume = {18},
@@ -105,7 +104,7 @@ author = {Gutierrez, A.P. and Ponti, L. and Neteler, M. and Stocchino, F. and Cu
 year = {2025},
 month = {04},
 pages = {},
-title = {Geospatial physiologically based demographic modeling and analysis of thirteen invasive species in Africa: why the biology matters},
+title = {Geospatial physiologically based demographic modeling and analysis of thirteen invasive species in {A}frica: why the biology matters},
 volume = {37},
 journal = {Environmental Sciences Europe},
 doi = {10.1186/s12302-025-01084-y}
