+++ paper.bib
@@ -111,7 +111,7 @@ doi = {10.1186/s12302-025-01084-y}
 }
 
 @article{g-y-n-e:1999,
-title = {A physiologically-based tritrophic metapopulation model of the African cassava food web},
+title = {A physiologically-based tritrophic metapopulation model of the {A}frican cassava food web},
 journal = {Ecological Modelling},
 volume = {123},
 number = {2},
