+++ docs/development/development_team.md
@@ -10,6 +10,7 @@ This is the current team who maintain a collective responsibility for the develo
 | -------------------- | -------------------- | -------------------- | ------ |
 | [Chiara Facciolà](https://github.com/fagiothree) | Mathematical Scientist, [IDEMS International](https://www.idems.international/) | <ul><li>Applied mathematical modelling</li><li>Examples of complex systems modelling</li></ul>  | 2024 + |
 | [Andrew Gutierrez](https://github.com/AndrewPGutierrez) | <ul><li>Professor Emeritus of Ecosystem Science, [UC Berkeley](https://www.berkeley.edu/)</li><li>CEO, [CASAS Global](https://casasglobal.org/)</li></ul>  | Biological and ecosystem modelling | 2023 + |
+| [Tim Hosgood](https://github.com/thosgood) | Research Scientist (Affiliate), [Topos Institute](https://topos.institute/) | Applied category theory and collaborative modelling | 2025 + |
 | [Georg Osang](https://github.com/geoo89) | Postdoctoral Impact Activation Fellow, [IDEMS International](https://www.idems.international/) | Code developer and maintainer | 2023 + |
 | [Luigi Ponti](https://github.com/luisponti) | <ul><li>Research scientist, [ENEA](https://www.enea.it/en/)</li><li>Research fellow, [CASAS Global](https://casasglobal.org/)</li></ul> | <ul><li>Biological and ecosystem modelling</li><li>Github structures and management</li></ul> | 2023 + |
 | [George Simmons](https://github.com/GJHSimmons) | Postdoctoral Impact Activation Fellow, [IDEMS International](https://www.idems.international/) | <ul><li>Code developer and maintainer</li><li>Applied category theory and collaborative modelling</li><li>Github structures and management</li></ul> | 2023 + |
