+++ citation.cff
@@ -1,0 +1,29 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+  - family-names: "Simmons"
+    given-names: "George"
+    orcid: "https://orcid.org/0000-0002-9076-4973"
+  - family-names: "Stern"
+    given-names: "David"
+  - family-names: "Osang"
+    given-names: "Georg"
+    orcid: "https://orcid.org/0000-0002-8882-5116"
+  - family-names: "Ponti"
+    given-names: "Luigi"
+    orcid: "https://orcid.org/0000-0003-4972-8265"
+  - family-names: "Gutierrez"
+    given-names: "Andrew Paul"
+    orcid: "https://orcid.org/0000-0001-7773-1715"
+  - family-names: "Facciolà"
+    given-names: "Chiara"
+    orcid: "https://orcid.org/0000-0001-8359-9300"
+  - family-names: "Hosgood"
+    given-names: "Timothy"
+    orcid: "https://orcid.org/0000-0003-4580-0741"
+title: "psymple"
+abstract: "Applied category-theoretic hybrid modelling of complex systems."
+license: GPL-3.0
+doi: 10.5281/zenodo.15237709
+url: "https://casasglobal-org.github.io/psymple"
+repository-code: "https://github.com/casasglobal-org/psymple"
