### 2025-05-04 #7364 @GJHSimmons
@prashjha

Thank you for the confirmation.

Here are the changes made:

- Paper updated. Here are the changes from the previous version: [Psymple JOSS review.docx](https://github.com/user-attachments/files/20029840/Psymple.JOSS.review.docx)
- Development credits updated in the [documentation](https://casasglobal-org.github.io/psymple/latest/development/development_team/)
- Zenodo updated with new version and authorship. Now version 1.0.2. [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.15338173.svg)](https://doi.org/10.5281/zenodo.15338173)

Thank you again!
