### 2025-05-05 #7364 @prashjha
@GJHSimmons, please update the tittle of the [zenodo archive](https://zenodo.org/records/15338173) to match the title of your JOSS paper.
