### 2025-05-06 bbc0ada
paper update
