+++ paper.bib
@@ -15,7 +15,7 @@ doi = {10.1016/B978-0-12-809633-8.02390-6}
   volume = {2},
   number = {6},
   pages = {783-800},
-  doi = {https://doi.org/10.1002/wcc.148},
+  doi = {10.1002/wcc.148},
   url = {https://wires.onlinelibrary.wiley.com/doi/abs/10.1002/wcc.148},
   eprint = {https://wires.onlinelibrary.wiley.com/doi/pdf/10.1002/wcc.148},
   abstract = {Abstract Earth System models (ESMs) are global climate models with the added capability to explicitly represent biogeochemical processes that interact with the physical climate and so alter its response to forcing such as that associated with human-caused emissions of greenhouse gases. Representing the global carbon cycle allows for feedbacks between the physical climate and the biological and chemical processes in the ocean and on land that take up some of the emitted carbon dioxide and so act to reduce warming. The sulfur cycle is also important in that both natural and human emissions of sulfur contribute to the production of sulfate aerosols which reflect incoming solar radiation (a direct cooling effect) and alter cloud properties (an indirect cooling effect). Other components such as ozone are also being incorporated into some ESMs. Evaluating the physical component of an ESM is becoming increasingly comprehensive and sophisticated, but the evaluation of the biogeochemical components suffer somewhat from a lack of comprehensive global-scale observational data. Nevertheless, such models provide valuable insight into climate variability and change, and the role of human activities and possible mitigation actions on future climate change. Internationally coordinated experiments are increasingly important in providing a multimodel ensemble of climate simulations, thereby taking advantage of some ‘cancellation of errors’ and allowing better quantification of uncertainty. WIREs Clim Change 2011, 2:783–800. doi: 10.1002/wcc.148 This article is categorized under: Climate Models and Modeling > Earth System Models},
@@ -225,7 +225,7 @@ doi = {10.1007/s10340-023-01610-y}
   year = {2005},
   note = {Integrating Ecology and Economics in Control Bioinvasions},
   issn = {0921-8009},
-  doi = {https://doi.org/10.1016/j.ecolecon.2004.05.007},
+  doi = {10.1016/j.ecolecon.2004.05.007},
   url = {https://www.sciencedirect.com/science/article/pii/S0921800904003106},
   author = {Gutierrez, A.P. and Regev, U.},
   keywords = {Economic–ecological theory, Energy flow, Invasive species},
@@ -335,20 +335,30 @@ title = {ILCYM-Insect life cycle modeling: software for developing temperature-b
   number = {3-4},
 }
 
+@book{guti:1996,
+	address = {New York},
+	title = {Applied {Population} {Ecology}: {A} {Supply}-{Demand} {Approach}},
+	isbn = {978-0-471-13586-9},
+	shorttitle = {Applied {Population} {Ecology}},
+	publisher = {John Wiley \& Sons},
+	author = {Gutierrez, A.P.},
+	year = {1996},
+}
+
 @article{guti:2025a,
   author = {Gutierrez, A.P. and Ponti, L. and Levi-Mourao, A. and Pons, X. and Cure, J.R. and Neteler, M. and Simmons, G.},
   title = {Stabilizing adaptation in an invasive species: {A}lfalfa weevil as a case study},
   journal = {Agricultural and Forest Entomology},
   year = {2025},
   keywords = {geographic distribution, invasive species, pest risk analysis, physiologically based demographic models, population dynamics, relative abundance, stabilizing selection, strains},
-  doi = {https://doi.org/10.1111/afe.12686},
+  doi = {10.1111/afe.12686},
   url = {https://resjournals.onlinelibrary.wiley.com/doi/abs/10.1111/afe.12686},
   eprint = {https://resjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/afe.12686},
[…]
