+++ paper.md
@@ -56,7 +56,7 @@ The development of `psymple` emerged from the complex system modelling requireme
 
 An alternative approach is mechanistic modelling, which instead of observational data use physiological data to capture the underlying mechanisms which drive species distribution, such as energy balance, population dynamics or response to climate [@ke-po:2009]. In contrast to correlative approaches, mechanistic SDMs decouple the physiology of a species from their native geography or climate, and allow SDMs in new geographic or climatic regimes to be created in the absence of observational data [@john:2019]. 
 
-An example mechanistic framework is physiologically-based demographic modelling (PBDM) [@guti:1996], which creates holistic ecosystem models based on the weather-driven biology of component species, allowing for predictive phenology, age- or mass-structured population dynamics, and geographic distribution assessments [@gu-po:2022a; @gu-po:2022b; @guti:2025a] for an overview and further references. With this approach, PBDM can account for tritrophic ecosystem interactions [@g-y-n-e:1999], or model the effects of climate change [@guti:2023].
+An example mechanistic framework is physiologically-based demographic modelling (PBDM) [@guti:1996], which creates holistic ecosystem models based on the weather-driven biology of component species, allowing for predictive phenology, age- or mass-structured population dynamics, and geographic distribution assessments [@gu-po:2022a; @gu-po:2022b; @guti:2025a]. With this approach, PBDM can account for tritrophic ecosystem interactions [@g-y-n-e:1999], or model the effects of climate change [@guti:2023].
 
 While correlative SDMs and mechanistic frameworks such as PBDM are today often regarded as conceptually distinct and largely unintegrated [@dorm:2012], their development can be traced back to early common roots [@fi-ni:1970; @guti:1974; @dw-go:1978]. A component of the PBDM framework is the use of physiological data to parametrise "biodemographic" functions capturing biophysical or biochemical mechanisms, such as the development, mortality and fecundity rates of a species in response to environmental variables [@po-gu:2023].
 
@@ -70,16 +70,16 @@ Models in `psymple` are built from arbitrary combinations of modular mechanistic
 
 # Description
 
-The workings of `psymple` are based on the ideas of the dynamic systems modelling package `AlgebraicJulia.jl` [@l-b-p-f:2024], in which modular objects, called *resource sharers*, containing ordinary differential equations (ODEs), are linked by wires representing the sharing of state variables across the dynamics equations of the system. The objects and wires form a formal diagram in applied category theory which can be interpreted as algebras over the operad of undirected wiring diagrams [@l-b-p-f:2022]. A compilation process maps this diagram to a system of ODEs by adding the right-hand side of all states connected to the same wiring system. 
+The workings of `psymple` are based on the ideas of the dynamic systems modelling package `AlgebraicJulia.jl` [@l-b-p-f:2024], in which modular objects, called *resource sharers*, containing ordinary differential equations (ODEs), are linked by wires representing the sharing of state variables across the dynamics equations of the system. In mathematical applied category theory, the objects and wires form a formal diagram which can be interpreted as algebras over the operad of undirected wiring diagrams [@l-b-p-f:2022]. A compilation process maps this diagram to a system of ODEs by adding the right-hand side of all states connected to the same wiring system. 
 
[…]
