+++ paper.md
@@ -104,6 +104,6 @@ For example, research is currently active in bioeconomic modelling, which extend
 
 # Acknowledgements
 
-The collaboration between IDEMS International and CASAS Global is enabled by the McKnight Foundation's [Global Collaboration for Resilient Food Systems](https://www.ccrp.org/), grant number $23-149$. The authors would like to thank the members of CASAS Global for their insights in developing the vision for a modern implementation of PBDM, and Brendan Fong and the rest of the Topos team in Berkeley for hosting a seminar that kick-started much of the shared vision of this work.
+The collaboration between IDEMS International and CASAS Global is enabled by the McKnight Foundation's [Global Collaboration for Resilient Food Systems](https://www.ccrp.org/), grant numbers $22-341$, $23-149$, and $24-124$. The authors would like to thank the members of CASAS Global for their insights in developing the vision for a modern implementation of PBDM, and Brendan Fong and the rest of the Topos team in Berkeley for hosting a seminar that kick-started much of the shared vision of this work.
 
 # References
