### 2025-05-06 #7364 @GJHSimmons
@prashjha

Thank you for getting back quickly with your comments.

- The zenodo title has been updated.
- The statement of need has been separated from a new section "background in ecological modelling". Hopefully this makes it clearer.
- I've updated the "description" section with a more detailed mathematical description and toy example which explains what is going on. The mirror code of the example could be added here, but that will add approximately another page. Personally I don't think this would add much, but let us know what you think. Instead, I've added a link to the documentation for more information.
- The remaining DOI error suggestion points to a book review, and not the book itself.
