### 2025-05-06 #7364 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
