### 2025-05-06 6403570
paper + bib update
