+++ paper.bib
@@ -189,7 +189,7 @@ doi = {10.4204/EPTCS.372.14}
 }
 
 @article{meur:2017,
-    title = {SymPy: symbolic computing in Python},
+    title = {SymPy: symbolic computing in {P}ython},
     author = {Meurer, Aaron and Smith, Christopher P. and Paprocki, Mateusz and \v{C}ert\'{i}k, Ond\v{r}ej and Kirpichev, Sergey B. and Rocklin, Matthew and Kumar, AMiT and Ivanov, Sergiu and Moore, Jason K. and Singh, Sartaj and Rathnayake, Thilina and Vig, Sean and Granger, Brian E. and Muller, Richard P. and Bonazzi, Francesco and Gupta, Harsh and Vats, Shivam and Johansson, Fredrik and Pedregosa, Fabian and Curry, Matthew J. and Terrel, Andy R. and Rou\v{c}ka, \v{S}t\v{e}p\'{a}n and Saboo, Ashutosh and Fernando, Isuru and Kulal, Sumith and Cimrman, Robert and Scopatz, Anthony},
     year = 2017,
     month = jan,
@@ -237,7 +237,7 @@ author = {Gutierrez, A.P. and Ponti, L. and Kranthi, K. and Baumgärtner, J. and
 year = {2020},
 month = {12},
 pages = {139},
-title = {Bio-economics of Indian hybrid Bt cotton and farmer suicides},
+title = {Bio-economics of {I}ndian hybrid {B}t cotton and farmer suicides},
 volume = {32},
 journal = {Environmental Sciences Europe},
 doi = {10.1186/s12302-020-00406-6}
