### 2025-05-06 #7364 @GJHSimmons
@danielskatz here are the changes made:

**Paper**
- Line 84/85 - mathematical applied category theory -> applied category theory
- Line 89: dx_i/dt -> dx_1/dt

**References**
- Line 196 indian -> Indian
- Line 240 python -> Python
