### 2025-05-07 #7364 @danielskatz
@GJHSimmons - I've suggested a bunch of changes in https://github.com/casasglobal-org/psymple/pull/71, all minor. Please merge this, or let me know what you disagree with, and then we can proceed to acceptance and publication.
