### 2025-05-07 27d5eda
Update CITATION.cff
