### 2025-05-07 #7364 @GJHSimmons
@danielskatz I've merged your suggestions with no disagreements - thank you for being so thorough.

Two final changes from me (already merged):
Line 103 (math display environment): (z,t,h(z,t,i_2)) -> g(z,t,h(z,t,i_2))
Line 109: VPO, FPO, and CPO objects -> VPO, FPO, and CPO instances (O already means object)
