### 2025-05-07 #7364 @danielskatz
Congratulations to @GJHSimmons (George Simmons) and co-authors on your publication!!

JOSS depends fully on volunteers and we couldn't do this without you
