### 2025-05-07 #7364 @GJHSimmons
Thank you @pitmonticone and @SunnyXu for your thorough reviews and comments, as well as your patience early on in the review when time was short to reply to your inputs.

Thank you @prashjha for overseeing the review, keeping us on track, and your helpful guidance and suggestions.

And finally @danielskatz thank you for your thorough inputs, as well as getting us over the line quickly at the end. We're very happy and proud to be part of the JOSS family and framework.
